Add more stages to the dependency resolver
The previous iteration of the resolver blindly checked incompatibilities
without checking if they are enabled. To fix this, it's now split into
a few more steps:

- dependency resolver now resolves what is and isn't enabled instead of
the extension loader (which used to do that after invoking the resolver)
- a preliminary stage runs, then it does implicit dependencies, then it
does incompatibility checks, then it repairs the dependency graph if it
broke from disabling incompatibilities

I haven't tested dependencies being removed from incompatibilities, or
stuck entries on the dependency graph, but it should theoretically work.
Please leave an issue if you encounter any problems or infinite loops.