darwin: make codesigning utils sync, properly hook them
I missed that hookUpdaterForPersistence (FKA persistAsar) merely hooks
into the updater, so we need to run it inside of the hooked emit method.
Those are synchronous, so our shelling out to codesign(1) needs to be
synchronous, too. This blocks the updater. Not sure if this can somehow
be async instead.