core: Support multiple Webpack scripts
The primary motivation for this is loading libdiscoreWasmFetch along
with web, as well as cleaning up Webpack patching code in general. We
have to support *multiple requires* existing, so the main require in
web gets marked with a special field. There's also multiple modules,
some with conflicting IDs(!!!), so I'm not sure how that behaves. I do
not know if patching modules in the other scripts works.

I haven't tested this very well, so feedback is appreciated! Browser
support isn't done yet, and noTrack will also need to be updated for
the libdiscore launch signature now that it's actually being evaluated.